+++ core-application/build.gradle
@@ -14,6 +14,7 @@ dependencies {
 
     implementation 'org.springframework:spring-tx'
     implementation 'org.springframework:spring-context'
+    implementation 'org.springframework.data:spring-data-commons'
     implementation 'com.fasterxml.jackson.core:jackson-annotations:2.15.2'
     implementation 'jakarta.validation:jakarta.validation-api'
