+++ core-domain/build.gradle
@@ -10,8 +10,6 @@ repositories {
 }
 
 dependencies {
-    implementation 'org.springframework.data:spring-data-commons'
-
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
     testImplementation 'org.junit.jupiter:junit-jupiter-params:5.9.2'
     testImplementation 'org.assertj:assertj-core:3.24.2'
